split data up into a Junction's Configuration and its metrics, saving and loading object works, field repopulation not impl yet
